+++ file.h
@@ -20,6 +20,6 @@ char *file_name(char *path);
 char **extract_ext(char *filename);
 
 // iterates over all files starting from a directory, can be recursive and call itself on subdirectories
-void iterate_files(const char *cur_dir, const char *search, short outer, short flags);
+void iterate_files(const char *cur_dir, const char *search, short outer, int flags);
 
 #endif //LOOK4_FILE_H
